+++ wkz/best_sections/generic.py
@@ -10,7 +10,7 @@ class GenericBestSection:
     kind: str
 
 
-def _activity_suitable_for_awards(activity) -> bool:
+def activity_suitable_for_awards(activity) -> bool:
     if activity.evaluates_for_awards is False or activity.sport.evaluates_for_awards is False:
         return False
     else: